(wait_grp,wait_task): changed type of loop index "i" from size_t to int to avoid signed/unsigned comparison

+++ trunk/ace/Thread_Manager.cpp
@@ -751,7 +751,7 @@ ACE_Thread_Manager::wait_grp (int grp_id)
   
   // Now to do the actual work
   int result = 0;
-  for (size_t i = 0; i < copy_count && result != -1; i++)
+  for (int i = 0; i < copy_count && result != -1; i++)
     if (ACE_Thread::join (copy_table[i].thr_handle_) == -1)
       result = -1;
   
@@ -854,7 +854,7 @@ ACE_Thread_Manager::wait_task (ACE_Task_Base *task)
   
   // Now to do the actual work
   int result = 0;
-  for (size_t i = 0; i < copy_count && result != -1; i++)
+  for (int i = 0; i < copy_count && result != -1; i++)
     if (ACE_Thread::join (copy_table[i].thr_handle_) == -1)
       result = -1;
